+++ src/main/java/janggi/domain/board/setup/BoardSetUp.java
@@ -1,7 +1,7 @@
 package janggi.domain.board.setup;
 
-import janggi.domain.coordinate.Point;
-import janggi.domain.piece.Piece;
+import janggi.domain.board.coordinate.Point;
+import janggi.domain.piece.unit.Piece;
 import janggi.domain.side.Side;
 import java.util.Map;
