+++ src/main/java/janggi/domain/board/setup/InSetUp.java
@@ -1,9 +1,9 @@
 package janggi.domain.board.setup;
 
-import janggi.domain.coordinate.Point;
-import janggi.domain.piece.Elephant;
-import janggi.domain.piece.Horse;
-import janggi.domain.piece.Piece;
+import janggi.domain.board.coordinate.Point;
+import janggi.domain.piece.unit.Elephant;
+import janggi.domain.piece.unit.Horse;
+import janggi.domain.piece.unit.Piece;
 import janggi.domain.side.Side;
 import java.util.HashMap;
 import java.util.Map;
