+++ src/main/java/janggi/domain/piece/Direction.java
@@ -1,4 +1,4 @@
-package janggi.domain.coordinate;
+package janggi.domain.piece;
 
 public enum Direction {
     NORTH(0, 1),
