+++ src/main/java/janggi/domain/piece/unit/Advisor.java
@@ -1,20 +1,21 @@
-package janggi.domain.piece;
+package janggi.domain.piece.unit;
 
-import janggi.domain.coordinate.Direction;
-import janggi.domain.coordinate.FixedPathStrategy;
-import janggi.domain.coordinate.Path;
-import janggi.domain.coordinate.PathStrategy;
-import janggi.domain.coordinate.Point;
+import janggi.domain.board.coordinate.FixedPathStrategy;
+import janggi.domain.board.coordinate.Path;
+import janggi.domain.board.coordinate.PathStrategy;
+import janggi.domain.board.coordinate.Point;
+import janggi.domain.piece.Direction;
+import janggi.domain.piece.PieceName;
 import janggi.domain.side.Side;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
-public class General extends Piece {
-    private static final PieceName NAME = PieceName.GENERAL;
+public class Advisor extends Piece {
+    private static final PieceName NAME = PieceName.ADVISOR;
     private static final PathStrategy DEFAULT_STRATEGY = new FixedPathStrategy();
 
-    public General(Side side) {
+    public Advisor(Side side) {
         super(NAME, side, DEFAULT_STRATEGY);
     }
 
@@ -36,6 +37,11 @@ public class General extends Piece {
 
     @Override
     protected Path cutPath(Path path, Map<Point, Piece> piecesOnPaths) {
-        return List.of();
+        return null;
+    }
+
+    @Override
+    protected boolean isValidPath(Path path, Map<Point, Piece> piecesOnPaths) {
+        return false;
     }
 }
