+++ src/main/java/janggi/domain/piece/unit/Empty.java
@@ -1,7 +1,8 @@
-package janggi.domain.piece;
+package janggi.domain.piece.unit;
 
-import janggi.domain.coordinate.Path;
-import janggi.domain.coordinate.Point;
+import janggi.domain.board.coordinate.Path;
+import janggi.domain.board.coordinate.Point;
+import janggi.domain.piece.PieceName;
 import janggi.domain.side.Side;
 import java.util.List;
 import java.util.Map;
