+++ src/main/java/janggi/domain/piece/unit/General.java
@@ -1,0 +1,42 @@
+package janggi.domain.piece.unit;
+
+import janggi.domain.board.coordinate.FixedPathStrategy;
+import janggi.domain.board.coordinate.Path;
+import janggi.domain.board.coordinate.PathStrategy;
+import janggi.domain.board.coordinate.Point;
+import janggi.domain.piece.Direction;
+import janggi.domain.piece.PieceName;
+import janggi.domain.side.Side;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+
+public class General extends Piece {
+    private static final PieceName NAME = PieceName.GENERAL;
+    private static final PathStrategy DEFAULT_STRATEGY = new FixedPathStrategy();
+
+    public General(Side side) {
+        super(NAME, side, DEFAULT_STRATEGY);
+    }
+
+    @Override
+    public List<Point> availablePoints(List<Path> paths, Map<Point, Piece> piecesOnPaths) {
+        return List.of();
+    }
+
+    @Override
+    public List<Path> path(Point from) {
+        List<Path> paths = new ArrayList<>();
+        for (Direction value : Direction.values()) {
+            Path path = convertToPath(List.of(value), from);
+            paths.add(path);
+        }
+
+        return paths;
+    }
+
+    @Override
+    protected Path cutPath(Path path, Map<Point, Piece> piecesOnPaths) {
+        return List.of();
+    }
+}
